Support for client side SSL. Large refactor of SSL system.
I've pushed most of the GnuTLS calls outside of liboi. The user is
responsible, now, for doing everything except the transport layer calls.
That is, the user should not access gnutls_transport_*  functions. Otherwise
the user is responsible for everything else including gnutls_global_init()
and such. The one interface that liboi provides is
oi_socket_set_secure_session().

The secure ping_pong test is passing 50% of the time. There is some bug in.