remove async->lock - rather have a global finished_lock
It is annoying to call the destrutor for the lock - it requires that each
oi_async object have a destructor. That, in turn, requires oi_file to have a
destructor. All for pthread_mutex_destroy, which in linux doesn't even do anything.
For now let's just have a global lock - it probably doesn't matter anyway.